librustc_data_structures: Unconfigure tests during normal build
