+++ src/librustc_data_structures/base_n.rs
@@ -3,6 +3,9 @@
 
 use std::str;
 
+#[cfg(test)]
+mod tests;
+
 pub const MAX_BASE: usize = 64;
 pub const ALPHANUMERIC_ONLY: usize = 62;
 pub const CASE_INSENSITIVE: usize = 36;
@@ -37,25 +40,4 @@ pub fn encode(n: u128, base: usize) -> String {
     let mut s = String::new();
     push_str(n, base, &mut s);
     s
-}
-
-#[test]
-fn test_encode() {
-    fn test(n: u128, base: usize) {
-        assert_eq!(Ok(n), u128::from_str_radix(&encode(n, base), base as u32));
-    }
-
-    for base in 2..37 {
-        test(0, base);
-        test(1, base);
-        test(35, base);
-        test(36, base);
-        test(37, base);
-        test(u64::max_value() as u128, base);
-        test(u128::max_value(), base);
-
-        for i in 0 .. 1_000 {
-            test(i * 983, base);
-        }
-    }
 }
