+++ src/librustc_data_structures/binary_search_util/tests.rs
