+++ src/librustc_data_structures/graph/dominators/tests.rs
@@ -1,4 +1,4 @@
-use super::super::test::TestGraph;
+use super::super::tests::TestGraph;
 
 use super::*;
