+++ src/librustc_data_structures/graph/scc/tests.rs
@@ -1,6 +1,4 @@
-#![cfg(test)]
-
-use crate::graph::test::TestGraph;
+use crate::graph::tests::TestGraph;
 use super::*;
 
 #[test]
