+++ src/librustc_data_structures/snapshot_map/mod.rs
@@ -4,7 +4,7 @@ use std::ops;
 use std::mem;
 
 #[cfg(test)]
-mod test;
+mod tests;
 
 pub struct SnapshotMap<K, V>
     where K: Hash + Clone + Eq
