+++ src/librustc_data_structures/snapshot_map/tests.rs
