+++ src/librustc_data_structures/tiny_list.rs
@@ -11,6 +11,9 @@
 //! If you expect to store more than 1 element in the common case, steer clear
 //! and use a `Vec<T>`, `Box<[T]>`, or a `SmallVec<T>`.
 
+#[cfg(test)]
+mod tests;
+
 #[derive(Clone, Hash, Debug, PartialEq)]
 pub struct TinyList<T: PartialEq> {
     head: Option<Element<T>>
@@ -116,141 +119,5 @@ impl<T: PartialEq> Element<T> {
         } else {
             false
         }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    extern crate test;
-    use test::Bencher;
-
-    #[test]
-    fn test_contains_and_insert() {
-        fn do_insert(i : u32) -> bool {
-            i % 2 == 0
-        }
-
-        let mut list = TinyList::new();
-
-        for i in 0 .. 10 {
-            for j in 0 .. i {
-                if do_insert(j) {
-                    assert!(list.contains(&j));
-                } else {
-                    assert!(!list.contains(&j));
-                }
-            }
-
-            assert!(!list.contains(&i));
-
-            if do_insert(i) {
-                list.insert(i);
-                assert!(list.contains(&i));
-            }
-        }
-    }
-
-    #[test]
-    fn test_remove_first() {
-        let mut list = TinyList::new();
-        list.insert(1);
-        list.insert(2);
-        list.insert(3);
-        list.insert(4);
-        assert_eq!(list.len(), 4);
-
-        assert!(list.remove(&4));
-        assert!(!list.contains(&4));
-
-        assert_eq!(list.len(), 3);
-        assert!(list.contains(&1));
-        assert!(list.contains(&2));
-        assert!(list.contains(&3));
-    }
-
-    #[test]
-    fn test_remove_last() {
-        let mut list = TinyList::new();
-        list.insert(1);
-        list.insert(2);
-        list.insert(3);
-        list.insert(4);
-        assert_eq!(list.len(), 4);
-
-        assert!(list.remove(&1));
-        assert!(!list.contains(&1));
-
-        assert_eq!(list.len(), 3);
-        assert!(list.contains(&2));
-        assert!(list.contains(&3));
-        assert!(list.contains(&4));
-    }
-
-    #[test]
-    fn test_remove_middle() {
-        let mut list = TinyList::new();
-        list.insert(1);
-        list.insert(2);
-        list.insert(3);
-        list.insert(4);
-        assert_eq!(list.len(), 4);
-
-        assert!(list.remove(&2));
-        assert!(!list.contains(&2));
-
-        assert_eq!(list.len(), 3);
-        assert!(list.contains(&1));
-        assert!(list.contains(&3));
-        assert!(list.contains(&4));
-    }
-
-    #[test]
-    fn test_remove_single() {
-        let mut list = TinyList::new();
-        list.insert(1);
-        assert_eq!(list.len(), 1);
-
-        assert!(list.remove(&1));
-        assert!(!list.contains(&1));
-
-        assert_eq!(list.len(), 0);
-    }
-
-    #[bench]
-    fn bench_insert_empty(b: &mut Bencher) {
-        b.iter(|| {
-            let mut list = TinyList::new();
-            list.insert(1);
-        })
-    }
-
-    #[bench]
-    fn bench_insert_one(b: &mut Bencher) {
-        b.iter(|| {
-            let mut list = TinyList::new_single(0);
-            list.insert(1);
-        })
-    }
-
-    #[bench]
-    fn bench_remove_empty(b: &mut Bencher) {
-        b.iter(|| {
-            TinyList::new().remove(&1)
-        });
-    }
-
-    #[bench]
-    fn bench_remove_unknown(b: &mut Bencher) {
-        b.iter(|| {
-            TinyList::new_single(0).remove(&1)
-        });
-    }
-
-    #[bench]
-    fn bench_remove_one(b: &mut Bencher) {
-        b.iter(|| {
-            TinyList::new_single(1).remove(&1)
-        });
     }
 }
